// generated/graphql.ts
export type Maybe<T> = T | null
// ...
export type Scalars = {
	ID: { input: string; output: string }
	String: { input: string; output: string }
	Boolean: { input: boolean; output: boolean }
	Int: { input: number; output: number }
	Float: { input: number; output: number }
	DateTime: { input: any; output: any }
	ObjectId: { input: any; output: any }
}
// ...
export type Content = {
	__typename?: "Content"
	/** The HTML version of the content. */
	html: Scalars["String"]["output"]
	/** The Markdown version of the content. */
	markdown: Scalars["String"]["output"]
	/** The text version from sanitized html content. HTML tags are stripped and only text is returned. */
	text: Scalars["String"]["output"]
}
// ...
export type DomainInfo = {
	__typename?: "DomainInfo"
	/** The domain of the publication. */
	domain?: Maybe<DomainStatus>
	/**
	 * The subdomain of the publication on hashnode.dev.
	 *
	 * It will redirect to you custom domain if it is present and ready.
	 */
	hashnodeSubdomain?: Maybe<Scalars["String"]["output"]>
	/** The www prefixed domain of the publication. Says if redirect to www domain is configured. */
	wwwPrefixedDomain?: Maybe<DomainStatus>
}
// ...
export type DraftConnection = Connection & {
	__typename?: "DraftConnection"
	/** A list of edges of drafts connection. */
	edges: Array<DraftEdge>
	/** Information to aid in pagination. */
	pageInfo: PageInfo
	/** The total number of documents in the connection. */
	totalDocuments: Scalars["Int"]["output"]
}
// ...
export type EmailImport = {
	__typename?: "EmailImport"
	/** Contains information about the current import example if it is in progress or has finished, date started, etc */
	currentImport?: Maybe<EmailCurrentImport>
}
// ...
export type Node = {
	/** The ID of the node. */
	id: Scalars["ID"]["output"]
}
// ...
export type Post = Node & {
	__typename?: "Post"
	/** Returns male and female audio url of the post. Available in case the Audioblog is enabled. */
	audioUrls?: Maybe<AudioUrls>
	/** Returns the user details of the author of the post. */
	author: User
	/**
	 * Flag to indicate if the post is bookmarked by the requesting user.
	 *
	 * Returns `false` if the user is not authenticated.
	 */
	bookmarked: Scalars["Boolean"]["output"]
	/** Brief is a short description of the post extracted from the content of the post. It's 250 characters long sanitized string. */
	brief: Scalars["String"]["output"]
	/** Canonical URL set by author in case of republished posts. */
	canonicalUrl?: Maybe<Scalars["String"]["output"]>
	/**
	 * Returns the user details of the co-authors of the post.
	 * Hashnode users can add up to 4 co-authors as collaborators to their posts.
	 * This functionality is limited to teams publication.
	 */
	coAuthors?: Maybe<Array<User>>
	/** List of users who have commented on the post. */
	commenters: PostCommenterConnection
	/** A list of comments on the post. */
	comments: PostCommentConnection
	/** Content of the post. Contains HTML and Markdown version of the post content. */
	content: Content
	/**
	 * A list of contributors of the post. Contributors are users who have commented or replied to the post.
	 * @deprecated Will be removed on 10th Oct 2023. Use `commenters` instead.
	 */
	contributors: Array<User>
	/** The cover image preference of the post. Contains cover image URL and other details. */
	coverImage?: Maybe<PostCoverImage>
	/** Unique ID to identify post, used internally by hashnode. */
	cuid?: Maybe<Scalars["String"]["output"]>
	/** Flag to indicate if the post is featured on Hashnode feed. */
	featured: Scalars["Boolean"]["output"]
	/** The date and time the post was featured. Used along with featured flag to determine if the post is featured. */
	featuredAt?: Maybe<Scalars["DateTime"]["output"]>
	/** Post feature-related fields. */
	features: PostFeatures
	/** A flag to indicate if the post contains LaTeX. Latex is used to write mathematical equations. */
	hasLatexInPost: Scalars["Boolean"]["output"]
	/** The ID of the post. Used to uniquely identify the post. */
	id: Scalars["ID"]["output"]
	/** Wether or not the post has automatically been published via RSS feed. */
	isAutoPublishedFromRSS: Scalars["Boolean"]["output"]
	/**
	 * Wether or not the authenticated user is following this post.
	 *
	 * Returns `null` if the user is not authenticated.
	 */
	isFollowed?: Maybe<Scalars["Boolean"]["output"]>
	/** A list of users who liked the post. */
	likedBy: PostLikerConnection
	/** OG meta-data of the post. Contains image url used in open graph meta tags. */
	ogMetaData?: Maybe<OpenGraphMetaData>
	/** Preference settings for the post. Contains information about if the post is pinned to blog, comments are disabled, etc. */
	preferences: PostPreferences
	/** The publication the post belongs to. */
	publication?: Maybe<Publication>
	/** The date and time the post was published. */
	publishedAt: Scalars["DateTime"]["output"]
	/** The number of hearts on the post. Shows how many users liked the post. */
	reactionCount: Scalars["Int"]["output"]
	/** The estimated time to read the post in minutes. */
	readTimeInMinutes: Scalars["Int"]["output"]
	/** The number of replies on the post. */
	replyCount: Scalars["Int"]["output"]
	/** The number of comments on the post. */
	responseCount: Scalars["Int"]["output"]
	/** SEO information of the post. Contains title and description used in meta tags. */
	seo?: Maybe<Seo>
	/** Information of the series the post belongs to. */
	series?: Maybe<Series>
	/** The slug of the post. Used as address of the post on blog. Example - https://johndoe.com/my-post-slug */
	slug: Scalars["String"]["output"]
	/** The subtitle of the post. Subtitle is a short description of the post which is also used in SEO if meta tags are not provided. */
	subtitle?: Maybe<Scalars["String"]["output"]>
	/** Returns list of tags added to the post. Contains tag id, name, slug, etc. */
	tags?: Maybe<Array<Tag>>
	/** The title of the post. */
	title: Scalars["String"]["output"]
	/** The date and time the post was last updated. */
	updatedAt?: Maybe<Scalars["DateTime"]["output"]>
	/** Complete URL of the post including the domain name. Example - https://johndoe.com/my-post-slug */
	url: Scalars["String"]["output"]
	/** The number of views on the post. Can be used to show the popularity of the post. */
	views: Scalars["Int"]["output"]
}
// ...
export type PostEdge = {
	__typename?: "PostEdge"
	/** A cursor for use in pagination. */
	cursor: Scalars["String"]["output"]
	/** The node holding the Post information */
	node: Post
}
// ...
export type Publication = Node & {
	__typename?: "Publication"
	/** The about section of the publication. */
	about?: Maybe<Content>
	/** The author who owns the publication. */
	author: User
	/** The canonical URL of the publication. */
	canonicalURL: Scalars["String"]["output"]
	/** The description of the publication, used in og:description meta tag. */
	descriptionSEO?: Maybe<Scalars["String"]["output"]>
	/** The title of the publication. Shown in blog home page. */
	displayTitle?: Maybe<Scalars["String"]["output"]>
	/** Domain information of the publication. */
	domainInfo: DomainInfo
	/** Returns the list of drafts in the publication. */
	drafts: DraftConnection
	/** Returns the publication's email imports, used with newsletter feature. */
	emailImport?: Maybe<EmailImport>
	/** The favicon of the publication. Used in browser tab. */
	favicon?: Maybe<Scalars["String"]["output"]>
	/** Object containing information about beta features enabled for the publication. */
	features: PublicationFeatures
	/** Total number of followers of the publication. */
	followersCount?: Maybe<Scalars["Int"]["output"]>
	/** Whether the publication has earned any badges or not. */
	hasBadges: Scalars["Boolean"]["output"]
	/** Color code of the header color of the publication. Used to style header of blog. */
	headerColor?: Maybe<Scalars["String"]["output"]>
	/** The ID of the publication. */
	id: Scalars["ID"]["output"]
	/** Summary of the contact information and information related to copyrights, usually used in German-speaking countries. */
	imprint?: Maybe<Scalars["String"]["output"]>
	/** The integrations connected to the publication. */
	integrations?: Maybe<PublicationIntegrations>
	/** Returns true if GitHub backup is configured and active and false otherwise. */
	isGitHubBackupEnabled: Scalars["Boolean"]["output"]
	/** A flag to indicate if the publication is using Headless CMS. This can be used to check if the post redirect needs authentication. */
	isHeadless: Scalars["Boolean"]["output"]
	/** True if the publication is a team publication and false otherwise. */
	isTeam: Scalars["Boolean"]["output"]
	/** Links to the publication's social media profiles. */
	links?: Maybe<PublicationLinks>
	/** The meta tags associated with the publication. */
	metaTags?: Maybe<Scalars["String"]["output"]>
	/** Information about the publication's Open Graph metadata i.e. image. */
	ogMetaData: OpenGraphMetaData
	/** Returns the pinned post of the publication. */
	pinnedPost?: Maybe<Post>
	/** Returns the post with the given slug. */
	post?: Maybe<Post>
	/** Returns the list of posts in the publication. */
	posts: PublicationPostConnection
	/** The publication preferences around layout, theme and other personalisations. */
	preferences: Preferences
	/** Configured redirection rules for the publication. */
	redirectionRules: Array<RedirectionRule>
	/** Returns the scheduled drafts of the publication. */
	scheduledDrafts: DraftConnection
	/** Returns series by slug in the publication. */
	series?: Maybe<Series>
	/** Returns the list of series in the publication. */
	seriesList: SeriesConnection
	/** Contains the publication's sponsorships information. */
	sponsorship?: Maybe<PublicationSponsorship>
	/** Returns the static page with the given slug. */
	staticPage?: Maybe<StaticPage>
	/** Returns a list of static pages in the publication. */
	staticPages: StaticPageConnection
	/** Returns the list of submitted drafts in the publication. */
	submittedDrafts: DraftConnection
	/**
	 * The title of the publication.
	 * Title is used as logo if logo is not provided.
	 */
	title: Scalars["String"]["output"]
	/** The domain of the publication. Used to access publication. Example https://johndoe.com */
	url: Scalars["String"]["output"]
	/** Determines the structure of the post URLs. */
	urlPattern: UrlPattern
}
// ...
export type PublicationFeatures = {
	__typename?: "PublicationFeatures"
	/** Audio player for blog posts. */
	audioBlog: AudioBlogFeature
	/** Individual styling for the publication. */
	customCSS: CustomCssFeature
	/** Newsletter feature for the publication which adds a `/newsletter` route for collecting subscribers and allows sending out newsletters. */
	newsletter: NewsletterFeature
	/** Show the read time for blog posts. */
	readTime: ReadTimeFeature
	/** Widget that shows up if a text on a blog post is selected. Allows for easy sharing or copying of the selected text. */
	textSelectionSharer: TextSelectionSharerFeature
	/** Show the view count for blog posts. */
	viewCount: ViewCountFeature
}
// ...
export type PublicationLinks = {
	__typename?: "PublicationLinks"
	/** Daily.dev URL of the publication. */
	dailydev?: Maybe<Scalars["String"]["output"]>
	/** GitHub URL of the publication. */
	github?: Maybe<Scalars["String"]["output"]>
	/** Hashnode profile of author of the publication. */
	hashnode?: Maybe<Scalars["String"]["output"]>
	/** Instagram URL of the publication. */
	instagram?: Maybe<Scalars["String"]["output"]>
	/** LinkedIn URL of the publication. */
	linkedin?: Maybe<Scalars["String"]["output"]>
	/** Mastodon URL of the publication. */
	mastodon?: Maybe<Scalars["String"]["output"]>
	/** Twitter URL of the publication. */
	twitter?: Maybe<Scalars["String"]["output"]>
	/** Website URL of the publication. */
	website?: Maybe<Scalars["String"]["output"]>
	/** YouTube URL of the publication. */
	youtube?: Maybe<Scalars["String"]["output"]>
}